+++ packages/formats/src/index.ts
@@ -2,3 +2,4 @@ export * from "./rgbeffects";
 export * from "./fseq";
 export * from "./xsq";
 export * from "./midi";
+export * from "./papagayo";